Coordinar productos con propiedades siat

+++ app/Filament/Resources/Products/Pages/EditProduct.php
@@ -23,7 +23,9 @@ class EditProduct extends EditRecord
         if(!$this->data['has_optical_properties']){
             /** @var OpticalProperty $op */
             $op = OpticalProperty::where('product_id', $this->record->id)->first();
-            $op->delete();
+            if($op) {
+                $op->delete();
+            }
         }
     }
 }
